requirements.txt
fastapi
uvicorn[standard]
neo4j
python-dotenv
supabase
